requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
python-docx>=1.1,<2.0
requests>=2.32,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
